requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
Pillow==10.2.0
reportlab==4.0.9
altair==5.2.0
scikit-learn==1.4.1
python-dateutil==2.9.0
jinja2==3.1.3

# YOLO (Ultralytics)
ultralytics==8.1.28
opencv-python==4.9.0.80

# Optional dependencies used internally by Altair
jsonschema==4.21.1
toolz==0.12.1


#local
ollama 